Fixed a bug from print_reminder_detail and print_event_detail functions
I added if-else statement that checks if there is any reminder that exists in a particular event. Also, added if-else statements for all the detail components from print_detail_events functionality as some events do not contain a detail component and hence, without checking if it exists or not, it would cause an error.